# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VeriFactu SaaS API - Sistema de Facturación AEAT
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /taxpayers/{taxpayerId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List uploaded documents
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UploadedDocument"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload document for OCR extraction
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Document uploaded and OCR started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document with OCR extraction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document with extraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/create-draft:
    post:
      operationId: createDraftFromDocument
      tags: [documents]
      summary: Create invoice draft from OCR extraction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Draft created from OCR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # Dashboard
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        name:
          type: string
        nif:
          type: ["string", "null"]
        nifType:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - name
        - country
        - isActive
        - createdAt
        - updatedAt
...
    InvoiceSeries:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        prefix:
          type: string
        name:
          type: string
        currentNumber:
          type: integer
        year:
          type: integer
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - prefix
        - name
        - currentNumber
        - year
        - isDefault
        - isActive
        - createdAt
...
    InvoiceLine:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        vatRate:
          type: number
        discount:
          type: number
        subtotal:
          type: number
        vatAmount:
          type: number
        total:
          type: number
        productId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - invoiceId
        - description
        - quantity
        - unitPrice
        - vatRate
        - discount
        - subtotal
        - vatAmount
        - total
        - sortOrder
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        seriesId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceType:
          type: string
          enum: [STANDARD, SIMPLIFIED, RECTIFICATION, CANCELLATION]
        status:
          type: string
          enum: [DRAFT, EMITTED, CANCELLED, RECTIFIED]
        issueDate:
          type: ["string", "null"]
          format: date
        operationDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: ["string", "null"]
          format: date
        subtotal:
          type: number
        vatAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        originSource:
          type: string
          enum: [MANUAL, AI_DRAFT]
        rectifiedInvoiceId:
          type: ["integer", "null"]
        cancellationReason:
          type: ["string", "null"]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLine"
        client:
          $ref: "#/components/schemas/Client"
        series:
          $ref: "#/components/schemas/InvoiceSeries"
        verifactuRecord:
          $ref: "#/components/schemas/VeriFactuRecord"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - invoiceType
        - status
        - subtotal
        - vatAmount
        - total
        - originSource
        - lines
        - createdAt
        - updatedAt
...
    VeriFactuRecord:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        taxpayerId:
          type: integer
        recordType:
          type: string
          enum: [ALTA, ANULACION, SUBSANACION]
        status:
          type: string
          enum: [PENDING, SUBMITTED, ACCEPTED, ACCEPTED_WITH_ERRORS, REJECTED, ERROR]
        hash:
          type: ["string", "null"]
        previousHash:
          type: ["string", "null"]
        qrUrl:
          type: ["string", "null"]
        xmlPayload:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        aeatResponse:
          type: ["string", "null"]
        aeatCsv:
          type: ["string", "null"]
        aeatErrorCode:
          type: ["string", "null"]
        aeatErrorMessage:
          type: ["string", "null"]
        retryCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceId
        - taxpayerId
        - recordType
        - status
        - retryCount
        - createdAt
        - updatedAt
...
    UploadedDocument:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        filename:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        ocrStatus:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        extraction:
          $ref: "#/components/schemas/OcrExtraction"
        linkedInvoiceId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - filename
        - fileUrl
        - fileType
        - ocrStatus
        - createdAt
    OcrExtraction:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        isInvoice:
          type: ["boolean", "null"]
        confidence:
          type: ["number", "null"]
        supplierName:
          type: ["string", "null"]
        supplierNif:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        clientNif:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        issueDate:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        vatAmount:
          type: ["number", "null"]
        vatRate:
          type: ["number", "null"]
        total:
          type: ["number", "null"]
        iban:
          type: ["string", "null"]
        fieldConfidences:
          type: object
          additionalProperties:
            type: number
      required:
        - id
        - documentId
    UploadDocumentBody:
      type: object
      properties:
        filename:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileBase64:
          type: ["string", "null"]
      required:
        - filename
        - fileUrl
        - fileType

    # Dashboard
...